src/filters: bring back "obsolete" filters
I simply can't imagine a more effective way to break all of our
production vpy scripts than to mark all those filters as obsolete.

Now that I have learnt the trick, let's kick it up a notch, and
rename `vapoursynth.core` as the next step in the competition to
break the most user scripts. Of course, better and more effective
ideas are welcome. Please do not hesitate to contact the chief API
breaker.

Signed-off-by: akarin <i@akarin.info>