src/cython/vapoursynth.pxd, vapoursynth.pyx: add Plugin._register_func
This provides a way for the user to define Python filter function
in a plugin namespace and it also allows overriding existing filters.

While this functionality is very generic, our intended use is to provide
compatibility wrapper for filters. For example,

```
def tcanny(**args):
    if 'gmmax' in args:
        args['scale'] = args['gmmax'] / 255.0
        del args['gmmax']
    return core.tcanny._TCanny(**args)
core.tcanny._register_func('TCanny', \
     core.tcanny.get_functions()['TCanny']+'gmmax:float:opt;', \
     None, tcanny, override=True)
```

As this is an advanced API, there is no safe guard in place, and you
can even modify the `core.std` namespace. Please be careful.

Signed-off-by: akarin <i@akarin.info>