Add more detail to core.__getattr__ error message
Clarify that an attribute not existing may be due to plugins not
being installed, since this error message often confuses new users.
The wording "forget to install a plugin" (as opposed to, say,
"are you missing a plugin") was chosen to make the error message imply
that this can be fixed by installing the plugin.